fix: let extension use SharedArrayBuffer in chrome
Two additional keys are required in the manifest for chrome
extensions to be able to use SharedArrayBuffer.

Slight issue though: the keys cause a warning in Firefox upon install,
because Firefox doesn't know these keys. But the extension
functions normally.

We may wanna consider having a system to produce different manifests
for different browsers in the future.