fix: update Video model, removing "image" & "url"
Video was inheriting part of its properties from Recommendation,
which made the validation fail since Recommendation has a url
and mandatory image and Video has not.

So I removed the dependency to Recommendation and instead listed
all the fields that are present in database except creatorId,
which you seem to have left out on purpose.