fix: replace video preview by picture cuz CORS
Problem was, as you know, chrome won't load certain external
resources because of the strict CORS policy we had to set.

I've spent some time researching this and there doesn't seem
to be a solution to embed a youtube video in such a strict
context.

Word has it that YouTube just doesn't know how to serve videos in
a way that wouldn't leak data to or from the embedding page :)

So I replaced the video iframe with the thumbnail for now.