fix: display of Avatar - closes #45
mui-org has added the behaviour we need in this commit:
https://github.com/mui-org/material-ui/commit/972b994ae3876c78cc8a7af56ec0a568ad9607e9

But this version is not released yet, so meanwhile I did somthing
ugly and copied the Avatar file from our node_modules,
very slightly modified this ugly transpiled JS, and included it here.

Since we don't have a diff, here are the changes I made:

I modified the require statements lines 24 and 26 because they were
relative and did not work with a relocated Avatar.js.

I added line 108:
image.crossOrigin = 'anonymous';

I added line 165:
crossOrigin: "anonymous"

This file should be deleted once we migrate to the next version of
material-ui.