fix: cross-origin image loading in a generic way
This is a bit nasty, but that's the best solution I could
think of at the moment.

The problem is that:

- if we send an "origin" header to a server that
  doesn't set an "access-control-allow-origin" header,
  (e.g. YouTube) then the image doesn't load

- if we don't send an "origin" header to a server that
  **does** set the "access-control-allow-origin" header,
  the image doesn't load either

We cannot, I think, know in advance what the server will reply,
so I made a component that tries both ways.

Maybe this could be done more elegantly by making a HEAD request
to the server instead.