chore: disable the CSS maps in production
The app was not building because of, I think, a bug inside
"optimize-css-assets-webpack-plugin".

It's hard to debug because there are so many layers of configuration
involved. This is a good example of why I do not like react-scripts.

The added environment variable tells that css plugin not to generate
source-maps.

The JS soure-maps are still generated.

Ideally, we should generate regular source-maps for the front-end,
and no source-maps or inline source-maps for the back-end.